Fix for slow time mouseover tooltip.
